moved pathfind related classes into samw package
