+++ src/main/java/at/hannibal2/skyhanni/features/misc/pathfind/AreaNode.kt
@@ -1,8 +1,8 @@
-package at.hannibal2.skyhanni.features.misc.navigation
+package at.hannibal2.skyhanni.features.misc.pathfind
 
 import at.hannibal2.skyhanni.data.model.GraphNode
 import at.hannibal2.skyhanni.data.model.GraphNodeTag
-import at.hannibal2.skyhanni.features.misc.IslandAreas.getAreaTag
+import at.hannibal2.skyhanni.features.misc.pathfind.IslandAreas.getAreaTag
 
 // A GraphNode that has an area tag, representing a area border.
 data class AreaNode(
